Disable statx for all builds. JB#50106
libstd implements statx using syscall and thus it will not work on
scratchbox2. Moreover as it is supported only by Linux 4.11 and newer
it's little use for us at the moment. It can be disabled and which means
that optional extra information is set to None. Software should handle
that gracefully without issues thanks to rust's type system.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com
Signed-off-by: Matti Kosola <matti.kosola@jolla.com>