Provide ENV controls to bypass some sb2 calls between fork exec
In threaded systems there can be deadlocks caused by calling malloc
after fork and before exec. Rust manages this but sb2 does not.

The following ENV variables are available:

SB2_RUST_EXECVP_SHIM

This executable is passed to the execvp call. The intended executable
is passed in the argv[] list. Typically set to "/usr/bin/env --". Used
wth SB2_RUST_USE_REAL_EXECVP which will not do path processing to exec
this binary but the binary will then exec the command provided via sb2
lookups.

SB2_RUST_USE_REAL_EXECVP
When set will cause the libc execvp() to be called and not the sb2
gate

SB2_RUST_USE_REAL_FN
This ensures that no sb2 calls to gates for dup2/chdir etc are made
between fork and exec

SB2_RUST_NO_SPAWNVP
When set the rust libc::spawnvp() call (which is gated) is bypassed
and the fallback fork/exec is used.

Note that some reworking of rust's handling of program/argv[0] was
needed to support the SHIM functionality

Signed-off-by: David Greaves <david.greaves@jolla.com>